Refactor code and rename to libvideolive
